Export SIMD helpers via bound functions on sdr.SamplesC64
Prior to this, there was no way for depending libraries to use the SIMD
optimized multiplication routines. Since exporting the raw SIMD library
is not something I want to support, two helpers have been exported on
the sdr.SamplesC64 type, one to multiply each I and Q value by a real,
and the second will do a complex multiplication/rotation.