The MultiWriter API was kludgey, refactoring
It's a lot of work to pull the SampleRate -- and it should be checked on
Write anyway. We'll do the check up on top and prevent runtime errors in
favor of a constructor error. This also reduces the argument count.

There are only a few MultiWriters around the codebase, so this is an
easy change to make right now.