Break the entire goddamn universe.
This change removes one of the worst design decisions I made early on,
having sample rate being stored as a `uint32` rather than `uint`. Why
32? This was due to the rtlsdr API being a uint32_t, which suddenly had
an impact on how the Pluto and HackRF driver were implemented.

I still think it's worth being a uint, not an int because checking for
a negative sample rate is just a big trap waiting to happen, and also
something that has no semantic sense to it at all.