Add in a new helper to test read/write loops
This only tests that the amount of data written in matches the data
read out. This isn't always going to be the case (e.g. downsample), but
this will at least make it easy to test the entire codepath, and do
multiple Write/Read calls to fill the target buffer up.

Tests still need to check the data is /valid/, but this can be helpful
in ensuring we don't have obviously broken readers.