Re-add the HackRF RX Path
Only hiccup - the HackRF uses int8, *not* uint8. D'oh! As a result, I've
had to add a new IQ sample format, and add in the full matrix of
conversions. This has no SIMD acceleration or anything like that, but
for now I think that's fine to add in HackRF support.

I've also taken this moment to add some test coverage on conversions.