Add in an experimental LimeSDR driver.
This code has some gnarly bugs - and most of what the LimeSDR can do is
not fully implemented. For example, using the antenna ports, or multiple
channels is not built out. I'm not sure how this behaves, and I have
little to no confidence in the emissions.

Be very very careful when using this -- always be sure to monitor for
your local laws, and ensure there are no emissions interfearing with
other users.

This SDR also supports f32 out, making this the first(!) c64 native
radio -- but that's not implemented (yet!!).