Stream: "Fix" buffer lengths.
This fell out of the strem fix. Now that we're back to ReadFull at the
Stream (and not Bytes) layer, we suddenly can't process chonks less than
the internal buffer size (since an EOF will result in a 0, EOF).

The ReadTransformer is pretty jacked up to start with, I think it may be
worth trying to change how /most/ of the reader streams work to fix this
broken bit.

For now, let's align to the internal boundary, but this is *not* good.