[uhd] overhaul the rx channel / read closer
This gets rid of the loopy mess of anon functions and other messy
allocation handling in favor of an explicit closer. this isn't without
faults, but it's a bit easier to read, and a bit less loopy.

i'm not sure of the C.free calls are in the right place (are they
held by uhd?), but it feels a lot better so far.