[UHD] Add in TX side driver support.
This is extremely tempermental on timing, i'm getting a lot of
underflows. I think this is not /entirely/ the SDR's fault, since writes
do (in fact) block, which would lead to some lag while it rebuilds a
huge buffer that is used for many tx_send calls.

Some buffering in the pipe would help smooth that out a bit.