Add in a new "Yikes" package, for bad things.
This refactors out some of the worst helpers I've written, stuff like
treating a C memory allocation as a Go byte slice (gross) to allow for
Go to mutate the C buffer.

Additional "yikes" may be placed in here as time goes on. Please don't
copy any of these out :)