[stream.Add]: Support I16 and I8 natively
This will help avoid a Convert to do a simple IQ addition. [2]uint8
isn't supported yet since it's slightly more complicated and I want to
test it. That has some implications for simple phased applications of
the KerberosSDR, but I don't think it's grave since actually /phasing/
the IQ samples is best done with C64, since implementing complex
multiplication for uint8 is kinda overkill. For now :)