[stream] Start work on "BufPipe2"
This is a new (and unstable) API that I'm going to try and replace the
old bufpipe implementation with. I'm not happy with the hacks that begin
to cascade out of the Pipe sementic changes, and I want the BufPipe to
try and behave better. Specifically, for Rx/Tx path, the existing
behavior is a real bummer.

Please don't use 'stream.BufPipe2' unless you're testing things out; I
need to get the behavior documented, tested, and then switch over
implementations.