[uhd] switch to a BufPipe2 (for now)
This will let the TX path exhaust the BufPipe before it stops
transmission (truncating packets). There's now a similar problem where
packets get garbled at the head, which I need to fix in a way that's
better than doing the 10 buffer retransmit. 2 seems to work for a
B200mini, but the B210 needs a bit more. Since startup is "expensive"
already, this is going to work to prove things out. The tail is the
bigger deal for now.