[stream] Add in 2D beamform math
This works by picking a synthetic point somewhere on the antenna plane,
and computing the phase difference pairwise to each Antenna in the
system. Applying those phases would create a signal as seen at the
synthetic point.

For the 1D version, the "synthetic point" was the same as the 0th
Antenna, so this is a more generic version of the 1D code. The 1D code
may even be refactored soon :)