[sdr] turn PipeReader/PipeWriter into an interface
Long term, I want to make all "Pipes" pluggable. This is a step in that
direction. I'll be moving sdr.Pipe, stream.BufPipe2, stream.RingBuffer
to a sdr.Pipe{Reader,Writer,ReadWriter} soon, in the hopes we don't have
to bake specific queue tactics into the I/O between chunks.

There's also a world in which all of this is redone to be zero-copy,
but, not yet.