[beamform] Phase offsets should prob be conj'd
This means this will "unwind" the phase offsets to re-center the sample
in-phase - rather than producing complex values at the phase offset. The
vaules here are almost always going to be used to apply rotations to
phasors, I can't imagine anyone really wants the unconj'd values when
using this interface.