[*]: Start to add support for the -tags=static flag
This injects a new pkg-config option to force the link to become
--static, which can be used along with -ldflags='-extldflags "-static"'
to build a static binary that includes the .a files for some of our C
dependencies.

This will be added to all the C dependencies we can, so that we can
support a universal 'static' build tag that's widely respected.