[lime]: Remove LimeSDR support
While I *love* my LimeSDR as a formfactor, it's never been supported
right, and has had some severe problems with sample rate that aren't
something that are easily debugged. Given it's a hassle and not used
in practice in my code anywhere, and EOL upstream, it's unlikely this
code will live on in a meaningful way. This lets me drop a large c
dep, which is going to be helpful when porting to new platforms.