[rtl/kerberos] Add a "CoherentSdr" interface
This removes StartCoherentRx from the kerberos.Sdr type, since there's
no easy way to pass in a Planner, which is needed for sync. This smooths
out the interface to match API with the uhd driver, by forcing the
planner into a further wrapper. This means the offset driver is even
further wrapped by using the CoherentSdr, which uses kerberos.Sdr.

At some point I'm going to unwind all this, but man, today's not the
day.