[rtl/kerberos] Fix 'Multiply' behavior
Before, we'd rotate by the Phase for each Reader for the 1th -> nth, but
the 0th would remain as-is, since it has no phase offset from itself.
The only trouble with that is that the Multiply code will cause a
conversion to C64, which means the sdr.Readers will have a mixed
SampleFormat, which screws up the decoded SampleFormat from the coherent
set.

This reads better and can be optimized on the backend, which is likely
smarter anyway.