[stream/multiply] Add in a SampleFormatI8 Multiply handler
This mirrors the SampleFormatU8 Multiply helper, and builds a similar
table, but with the signed int8 format used by the HackRF and Ettus USRP
line. They both build up a table when a new constant is set, which is of
size INT16_MAX (aka 0xFFFF aka 65535). At 2Msps, that's 0.032 seconds of
IQ.