[pluto] Do over/underrun checking
This.... may not be documented. I found this deep in some docs a long
time ago but never put it on the hotpath. I think that was a mistake.
The UHD driver has been really helpful with debugging performance
issues. Dropped samples are also a LOT more annoying now that I'm doing
digital modes like OFDM and not FM or BPSK.

I'd like to re-add this to the hotpath; we'll see how this works. Maybe
this'll get an option later on.