[RingBuffer] Fix up Ring behavior ... again
This is the like 30th time this fscking thing has been the source of
headaches. The logic on Read blocks was flipped (doh!), and would
block forever if the Ring was closed. This cleans up both, but I have a
sneaking hunch this has even more bugs hidden in the dark waiting for
me.