[fft] Overhaul top-level fft.* helper API
The "Result" type was designed to be used as part of the FFT API,
perhaps even in fft.Plan, but that never materialized since it was just
too cumbersome. Since it was a bit unweildly to transform your frequency
domain result into a fft.Result, I wrote some helpers which internally
just created a Result, and allowed for method invocation but without
dealing with the struct nearly at all.

Of course, now some passed params became meaningless, and most notibly,
it required passing the whole frequency slice. The intent here was good
(it's creating a Result, and things like Shift can mutuate it); but what
it's done in practice was to force allocations. I've even written code
like 'fft.BinByFreq(make([]complex64, num), ...)' which is just
terrible.

I've flipped this so the helpers are now the implementation, and the
Result wrapper allows for easy invocation of those helpers as needed. At
some point I may even just deprecate and remove the whole Result type.